# ficha9/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Ficha 8 API",
        "description": "Ficha 8 API Information",
        "contact": {
            "name": "TPSI-DWB"
        },
        "servers": [
            "http://localhost:3000"
        ]
    },
    "paths": {
        "/persons": {
            "post": {
                "tags": [
                    "person"
                ],
                "summary": "Create a person",
                "description": "This can only be done by the logged in user.",
                "operationID": "createPerson",
                "produces": [
                    "aplication/xml",
                    "aplication/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created person object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Persons"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "sucessful operation"
                    }
                }
            }
        },
        "/persons/{id}": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "Get person by ID",
                "description": "Get person by ID",
                "operationID": "getPersonById",
                "produces": [
                    "aplication/xml",
                    "aplication/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "body",
                        "description": "The id that needs to be fetched. Use id 1 for testing",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Persons"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "sucessful operation"
                    }
                }
            }
        },
        "definitions": {
            "Persons": {
                "type": "object",
                "proprieties": {
                    "id": {
                        "type": "integer",
                        "x-primary-key": true
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            }
        }
    }
}